pyproject.toml
[tool.poetry]
name = "regen-analytics"
version = "0.1.0"
description = ""
authors = ["Aaron Craelius <aaronc@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
python-dotenv = "^0.18.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
